requirements.txt
about-time==4.2.1
alive-progress==3.2.0
appdirs==1.4.4
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
cssselect==1.3.0
fake-useragent==2.1.0
feedparser==6.0.11
filelock==3.18.0
fsspec==2025.3.0
grapheme==0.6.0
greenlet==3.1.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
lxml==5.3.1
lxml_html_clean==0.4.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parse==1.20.2
psycopg2-binary==2.9.10
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.39
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
w3lib==2.3.1
websockets==10.4
xmltodict==0.14.2
yahoo-fin==0.8.9.1
zipp==3.21.0
